Create LoginForm index.php
Write one PHP file that will output a form containing 3 fields: username, password and “Remember me!” checkbox. Upon submission of the form, the code should check against MySQL table to see whether the username/password pair was correct. If so, display a welcome message and save a cookie that identifies the user to the server. On further visits to the page, the user should appear logged in, even if the browser has been closed; If not, display the message “Invalid username or password” followed by the same login form. The output should be one of three options:

The login form.
The invalid message and the login form, if failed login.
The welcome message containing a “Log Out” button, if successful login.

When user clicks “Log out” button, remove the cookie so that the user is logged out. Clicking the button should redirect the user to the same page, which now shows a login form.